refactor: extract shared OAuthClientForm component
