+++ src/app/[orgId]/settings/oauth-clients/[clientId]/page.tsx
@@ -3,10 +3,8 @@
 import {
     SettingsContainer,
     SettingsSection,
-    SettingsSectionBody,
     SettingsSectionDescription,
     SettingsSectionFooter,
-    SettingsSectionForm,
     SettingsSectionHeader,
     SettingsSectionTitle
 } from "@app/components/Settings";
@@ -17,10 +15,7 @@ import { createApiClient, formatAxiosError } from "@app/lib/api";
 import { useEnvContext } from "@app/hooks/useEnvContext";
 import { toast } from "@app/hooks/useToast";
 import { Button } from "@app/components/ui/button";
-import { Input } from "@app/components/ui/input";
 import { Label } from "@app/components/ui/label";
-import { Checkbox } from "@app/components/ui/checkbox";
-import { Switch } from "@app/components/ui/switch";
 import {
     Credenza,
     CredenzaBody,
@@ -34,9 +29,9 @@ import {
 import CopyTextBox from "@app/components/CopyTextBox";
 import ConfirmDeleteDialog from "@app/components/ConfirmDeleteDialog";
 import { useTranslations } from "next-intl";
-import ClientAvatar from "@app/components/ClientAvatar";
 import { OAuthClient } from "../types";
 import { ResponseT } from "@server/types/Response";
+import OAuthClientForm, { type OAuthClientFormData } from "../OAuthClientForm";
 
 function parseScopes(scopeString: string): Set<string> {
     return new Set(
@@ -59,22 +54,11 @@ export default function EditOAuthClientPage() {
 
     const [loading, setLoading] = useState(true);
     const [saving, setSaving] = useState(false);
-    const [clientName, setClientName] = useState("");
-    const [clientUri, setClientUri] = useState("");
-    const [logoUri, setLogoUri] = useState("");
-    const [backchannelLogoutUri, setBackchannelLogoutUri] = useState("");
-    const [postLogoutRedirectUris, setPostLogoutRedirectUris] = useState<string[]>([""]);
-    const [redirectUris, setRedirectUris] = useState<string[]>([""]);
-    const [scopeProfile, setScopeProfile] = useState(true);
-    const [scopeEmail, setScopeEmail] = useState(true);
-    const [scopeGroups, setScopeGroups] = useState(false);
-    const [pkceRequired, setPkceRequired] = useState(true);
-    const [enabled, setEnabled] = useState(true);
+    const [client, setClient] = useState<OAuthClient | null>(null);
     const [rotatedSecret, setRotatedSecret] = useState<{
         clientId: string;
         clientSecret: string;
     } | null>(null);
-    const [showOptional, setShowOptional] = useState(false);
     const [isRotateModalOpen, setIsRotateModalOpen] = useState(false);
     const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
 
@@ -90,27 +74,7 @@ export default function EditOAuthClientPage() {
                 const res = await api.get<ResponseT<{ client: OAuthClient }>>(
                     `/org/${orgId}/oauth-clients/${clientId}`
                 );
-                const client = res.data.data!.client;
-                const clientScopes = parseScopes(client.scopes);
-
-                setClientName(client.clientName);
-                setClientUri(client.clientUri || "");
-                setLogoUri(client.logoUri || "");
-                setBackchannelLogoutUri(client.backchannelLogoutUri || "");
-                const postLogoutUris = client.postLogoutRedirectUris ?? [];
-                setPostLogoutRedirectUris(
-                    postLogoutUris.length > 0 ? postLogoutUris : [""]
-                );
-                setRedirectUris(
-                    client.redirectUris.length > 0
-                        ? client.redirectUris
-                        : [""]
-                );
-                setScopeProfile(clientScopes.has("profile"));
-                setScopeEmail(clientScopes.has("email"));
-                setScopeGroups(clientScopes.has("groups"));
-                setPkceRequired(client.pkceRequired);
-                setEnabled(client.enabled);
+                setClient(res.data.data!.client);
             } catch (error) {
                 toast({
                     variant: "destructive",
@@ -125,77 +89,21 @@ export default function EditOAuthClientPage() {
         load();
     }, [orgId, clientId]);
 
-    function updateRedirectUri(index: number, value: string) {
-        setRedirectUris((prev) =>
-            prev.map((item, i) => (i === index ? value : item))
-        );
-    }
-
-    function removeRedirectUri(index: number) {
-        setRedirectUris((prev) => prev.filter((_, i) => i !== index));
-    }
-
-    function updatePostLogoutRedirectUri(index: number, value: string) {
-        setPostLogoutRedirectUris((prev) =>
-            prev.map((item, i) => (i === index ? value : item))
-        );
-    }
-
-    function removePostLogoutRedirectUri(index: number) {
-        setPostLogoutRedirectUris((prev) => prev.filter((_, i) => i !== index));
-    }
-
-    async function saveClient() {
-        const cleanedRedirectUris = redirectUris
-            .map((item) => item.trim())
-            .filter((item) => item.length > 0);
-
-        if (!clientName.trim()) {
-            toast({
-                variant: "destructive",
-                title: t("oauthClientNameRequired")
-            });
-            return;
-        }
-
-        if (cleanedRedirectUris.length === 0) {
-            toast({
-                variant: "destructive",
-                title: t("oauthClientRedirectUrisRequired")
-            });
-            return;
-        }
-
+    async function handleSave(data: OAuthClientFormData) {
         setSaving(true);
-
         try {
-            const scopes = ["openid"];
-            if (scopeProfile) {
-                scopes.push("profile");
-            }
-            if (scopeEmail) {
-                scopes.push("email");
-            }
-            if (scopeGroups) {
-                scopes.push("groups");
-            }
-
-            const cleanedPostLogoutRedirectUris = postLogoutRedirectUris
-                .map((item) => item.trim())
-                .filter((item) => item.length > 0);
-
             await api.patch(`/org/${orgId}/oauth-clients/${clientId}`, {
-                clientName: clientName.trim(),
-                redirectUris: cleanedRedirectUris,
-                clientUri: clientUri.trim() || null,
-                logoUri: logoUri.trim() || null,
-                backchannelLogoutUri: backchannelLogoutUri.trim() || null,
-                postLogoutRedirectUris: cleanedPostLogoutRedirectUris.length > 0
-                    ? cleanedPostLogoutRedirectUris
+                clientName: data.clientName,
+                redirectUris: data.redirectUris,
+                clientUri: data.clientUri || null,
+                logoUri: data.logoUri || null,
+                backchannelLogoutUri: data.backchannelLogoutUri || null,
+                postLogoutRedirectUris: data.postLogoutRedirectUris.length > 0
+                    ? data.postLogoutRedirectUris
                     : null,
-                scopes,
-                pkceRequired,
-                enabled
+                scopes: data.scopes,
+                pkceRequired: data.pkceRequired,
+                enabled: data.enabled
             });
 
             toast({
@@ -250,7 +158,7 @@ export default function EditOAuthClientPage() {
         }
     }
 
-    if (loading) {
+    if (loading || !client) {
         return (
             <>
                 <SettingsHeaderTitle
@@ -263,6 +171,9 @@ export default function EditOAuthClientPage() {
             </>
         );
     }
+
+    const clientScopes = parseScopes(client.scopes);
+    const postLogoutUris = client.postLogoutRedirectUris ?? [];
 
     return (
         <>
@@ -364,281 +275,35 @@ export default function EditOAuthClientPage() {
                 }
                 buttonText={t("oauthClientDeleteButton")}
                 onConfirm={deleteClient}
-                string={clientName}
+                string={client.clientName}
                 title={t("oauthClientDeleteButton")}
             />
 
             <SettingsContainer>
-                <SettingsSection>
-                    <div className="flex items-center justify-between pb-4 mb-4">
-                        <div className="flex items-center gap-3">
-                            <ClientAvatar name={clientName} logoUri={logoUri} />
-                            <span className="font-medium text-lg truncate">
-                                {clientName}
-                            </span>
-                        </div>
-                        <Button
-                            variant="ghost"
-                            size="sm"
-                            onClick={() => setShowOptional((prev) => !prev)}
-                        >
-                            {showOptional
-                                ? t("oauthClientHideOptionalFields")
-                                : t("oauthClientShowOptionalFields")}
-                        </Button>
-                    </div>
-                    <SettingsSectionBody>
-                        <SettingsSectionForm>
-                            <div className="space-y-2">
-                                <Label>
-                                    {t("oauthClientIdHeader")}
-                                </Label>
-                                <CopyTextBox text={clientId} />
-                            </div>
-
-                            <div className="space-y-2">
-                                <Label htmlFor="client-name">
-                                    {t("oauthClientNameLabel")}
-                                </Label>
-                                <Input
-                                    id="client-name"
-                                    value={clientName}
-                                    onChange={(event) =>
-                                        setClientName(event.target.value)
-                                    }
-                                />
-                            </div>
-
-                            {(showOptional || clientUri.trim()) && (
-                                <div className="space-y-2">
-                                    <Label htmlFor="client-uri">
-                                        {t("oauthClientHomepageLabel")} <span className="text-muted-foreground font-normal">{t("optional")}</span>
-                                    </Label>
-                                    <Input
-                                        id="client-uri"
-                                        value={clientUri}
-                                        onChange={(event) =>
-                                            setClientUri(event.target.value)
-                                        }
-                                    />
-                                </div>
-                            )}
-
-                            {(showOptional || logoUri.trim()) && (
-                                <div className="space-y-2">
-                                    <Label htmlFor="logo-uri">
-                                        {t("oauthClientLogoLabel")} <span className="text-muted-foreground font-normal">{t("optional")}</span>
-                                    </Label>
-                                    <Input
-                                        id="logo-uri"
-                                        value={logoUri}
-                                        onChange={(event) =>
-                                            setLogoUri(event.target.value)
-                                        }
-                                    />
-                                </div>
-                            )}
-
-                            {(showOptional || backchannelLogoutUri.trim()) && (
-                                <div className="space-y-2">
-                                    <Label htmlFor="backchannel-logout-uri">
-                                        {t("oauthClientBackchannelLogoutUriLabel")} <span className="text-muted-foreground font-normal">{t("optional")}</span>
-                                    </Label>
-                                    <Input
-                                        id="backchannel-logout-uri"
-                                        value={backchannelLogoutUri}
-                                        onChange={(event) =>
-                                            setBackchannelLogoutUri(event.target.value)
-                                        }
-                                    />
-                                </div>
-                            )}
-
-                            {(showOptional || postLogoutRedirectUris.some(u => u.trim())) && (
-                                <div className="space-y-2">
-                                    <Label>
-                                        {t("oauthClientPostLogoutRedirectUrisLabel")} <span className="text-muted-foreground font-normal">{t("optional")}</span>
-                                    </Label>
-                                    <div className="space-y-3">
-                                        {postLogoutRedirectUris.map((uri, index) => (
-                                            <div key={index} className="flex gap-2">
-                                                <Input
-                                                    value={uri}
-                                                    onChange={(event) =>
-                                                        updatePostLogoutRedirectUri(
-                                                            index,
-                                                            event.target.value
-                                                        )
-                                                    }
-                                                    placeholder={t(
-                                                        "oauthClientPostLogoutRedirectUriPlaceholder"
-                                                    )}
-                                                />
-                                                {postLogoutRedirectUris.length > 1 && (
-                                                    <Button
-                                                        variant="outline"
-                                                        onClick={() =>
-                                                            removePostLogoutRedirectUri(index)
-                                                        }
-                                                    >
-                                                        {t("oauthClientRemovePostLogoutRedirectUri")}
-                                                    </Button>
-                                                )}
-                                            </div>
-                                        ))}
-                                        <Button
-                                            variant="outline"
-                                            size="sm"
-                                            onClick={() =>
-                                                setPostLogoutRedirectUris((prev) => [
-                                                    ...prev,
-                                                    ""
-                                                ])
-                                            }
-                                        >
-                                            {t("oauthClientAddPostLogoutRedirectUri")}
-                                        </Button>
-                                    </div>
-                                </div>
-                            )}
-
-                            <div className="space-y-2 pt-2">
-                                <Label>
-                                    {t("oauthClientRedirectUrisLabel")}
-                                </Label>
-                                <div className="space-y-3">
-                                    {redirectUris.map((redirectUri, index) => (
-                                        <div key={index} className="flex gap-2">
-                                            <Input
-                                                value={redirectUri}
-                                                onChange={(event) =>
-                                                    updateRedirectUri(
-                                                        index,
-                                                        event.target.value
-                                                    )
-                                                }
-                                                placeholder={t(
-                                                    "oauthClientRedirectUriPlaceholder"
-                                                )}
-                                            />
-                                            {redirectUris.length > 1 && (
-                                                <Button
-                                                    variant="outline"
-                                                    onClick={() =>
-                                                        removeRedirectUri(index)
-                                                    }
-                                                >
-                                                    {t(
-                                                        "oauthClientRemoveRedirectUri"
-                                                    )}
-                                                </Button>
-                                            )}
-                                        </div>
-                                    ))}
-                                    <Button
-                                        variant="outline"
-                                        size="sm"
-                                        onClick={() =>
-                                            setRedirectUris((prev) => [
-                                                ...prev,
-                                                ""
-                                            ])
-                                        }
-                                    >
-                                        {t("oauthClientAddRedirectUri")}
-                                    </Button>
-                                </div>
-                            </div>
-
-                            <div className="space-y-2 pt-2">
-                                <Label>
-                                    {t("oauthClientScopesLabel")}
-                                </Label>
-                                <p className="text-xs text-muted-foreground">
-                                    {t("oauthClientOpenidAlwaysEnabled")}
-                                </p>
-                                <div className="space-y-3 pt-1">
-                                    <div className="flex items-center gap-2">
-                                        <Checkbox
-                                            checked={scopeProfile}
-                                            onCheckedChange={(value) =>
-                                                setScopeProfile(value === true)
-                                            }
-                                            id="scope-profile"
-                                        />
-                                        <Label htmlFor="scope-profile">
-                                            {t("oauthClientScopeProfile")}
-                                        </Label>
-                                    </div>
-                                    <div className="flex items-center gap-2">
-                                        <Checkbox
-                                            checked={scopeEmail}
-                                            onCheckedChange={(value) =>
-                                                setScopeEmail(value === true)
-                                            }
-                                            id="scope-email"
-                                        />
-                                        <Label htmlFor="scope-email">
-                                            {t("oauthClientScopeEmail")}
-                                        </Label>
-                                    </div>
-                                    <div className="flex items-center gap-2">
-                                        <Checkbox
-                                            checked={scopeGroups}
-                                            onCheckedChange={(value) =>
-                                                setScopeGroups(value === true)
-                                            }
-                                            id="scope-groups"
-                                        />
-                                        <Label htmlFor="scope-groups">
-                                            {t("oauthClientScopeGroups")}
-                                        </Label>
-                                    </div>
-                                </div>
-                            </div>
-
-                            <div className="space-y-4 pt-2">
-                                <Label>
-                                    {t("oauthClientOptionsTitle")}
-                                </Label>
-                                <div className="flex items-center justify-between">
-                                    <Label htmlFor="pkce-required" className="font-normal">
-                                        {t("oauthClientRequirePkceLabel")}
-                                    </Label>
-                                    <Switch
-                                        id="pkce-required"
-                                        checked={pkceRequired}
-                                        onCheckedChange={setPkceRequired}
-                                    />
-                                </div>
-
-                                <div className="flex items-center justify-between">
-                                    <Label htmlFor="client-enabled" className="font-normal">
-                                        {t("oauthClientEnabledLabel")}
-                                    </Label>
-                                    <Switch
-                                        id="client-enabled"
-                                        checked={enabled}
-                                        onCheckedChange={setEnabled}
-                                    />
-                                </div>
-                            </div>
-                        </SettingsSectionForm>
-                    </SettingsSectionBody>
-
-                    <SettingsSectionFooter>
+                <OAuthClientForm
+                    initialValues={{
+                        clientName: client.clientName,
+                        clientUri: client.clientUri || "",
+                        logoUri: client.logoUri || "",
+                        backchannelLogoutUri: client.backchannelLogoutUri || "",
+                        postLogoutRedirectUris: postLogoutUris.length > 0 ? postLogoutUris : [""],
+                        redirectUris: client.redirectUris.length > 0 ? client.redirectUris : [""],
+                        scopeProfile: clientScopes.has("profile"),
+                        scopeEmail: clientScopes.has("email"),
+                        scopeGroups: clientScopes.has("groups"),
+                        pkceRequired: client.pkceRequired,
+                        enabled: client.enabled
+                    }}
+                    clientId={clientId}
+                    onSubmit={handleSave}
+                    submitting={saving}
+                    submitLabel={t("saveChanges")}
+                    footerExtra={
                         <Button variant="outline" onClick={() => setIsRotateModalOpen(true)}>
                             {t("oauthClientRotateButton")}
                         </Button>
-                        <Button
-                            onClick={saveClient}
-                            disabled={saving}
-                            loading={saving}
-                        >
-                            {t("saveChanges")}
-                        </Button>
-                    </SettingsSectionFooter>
-                </SettingsSection>
+                    }
+                />
 
                 <SettingsSection>
                     <SettingsSectionHeader>
